+++ src/Epidemic/Types/Newick.hs
@@ -2,17 +2,13 @@
 
 module Epidemic.Types.Newick where
 
-import qualified Data.Aeson as Json
-import qualified Data.ByteString as B
 import qualified Data.ByteString.Builder as BBuilder
 import qualified Data.List as List
 import qualified Data.Vector as V
-import Epidemic.Types.Parameter
 import Epidemic.Types.Observations
 import Epidemic.Types.Events
 import Epidemic.Types.Population
 import Epidemic.Types.Time
-import GHC.Generics
 
 -- | Class of types that can be expressed in Newick format.
 class Newick t
@@ -51,7 +47,7 @@ instance Newick ReconstructedTree where
           IndividualSample {..} ->
             if indSampSeq
               then Just
-                     ( (personByteString indSampPerson) <>
+                     ( personByteString indSampPerson <>
                        colonBuilder <> branchLength t indSampTime
                      , [e])
               else Nothing