+++ src/Epidemic/Types/Observations.hs
@@ -9,20 +9,22 @@ module Epidemic.Types.Observations
   , pointProcessEvents
   , reconstructedTreeEvents
   , observedEvents
+  , aggregated
   ) where
 
-import Control.Monad (liftM)
 import qualified Data.Aeson as Json
-import qualified Data.ByteString.Builder as BBuilder
 import qualified Data.List as List
-import qualified Data.Vector as V
 import Epidemic.Types.Events
   ( EpidemicEvent(..)
   , EpidemicTree(..)
   , maybeEpidemicTree
   )
-import Epidemic.Types.Time (TimeDelta(..), timeDelta)
-import Epidemic.Types.Population (People(..), personByteString)
+import Epidemic.Types.Population (asPeople)
+import Epidemic.Types.Time
+  ( TimeInterval(..)
+  , TimeStamp(..)
+  , inInterval
+  )
 import GHC.Generics
 
 -- | A wrapper for an 'EpidemicEvent' to indicate that this is an even that was
@@ -34,6 +36,9 @@ newtype Observation =
 instance Json.FromJSON Observation
 
 instance Json.ToJSON Observation
+
+instance TimeStamp Observation where
+  absTime (Observation ee) = absTime ee
 
 -- | A representation of the events that can be observed in an epidemic but
 -- which are not included in the reconstructed tree, ie the unsequenced
@@ -47,8 +52,10 @@ pointProcessEvents :: EpidemicTree -> PointProcessEvents
 pointProcessEvents Shoot {} = PointProcessEvents []
 pointProcessEvents (Leaf e) =
   case e of
-    IndividualSample {..} -> PointProcessEvents $ if not indSampSeq then [Observation e] else []
-    PopulationSample {..} -> PointProcessEvents $ if not popSampSeq then [Observation e] else []
+    IndividualSample {..} ->
+      PointProcessEvents [Observation e | not indSampSeq]
+    PopulationSample {..} ->
+      PointProcessEvents [Observation e | not popSampSeq]
     _ -> PointProcessEvents []
 pointProcessEvents (Branch _ lt rt) =
   let (PointProcessEvents lEs) = pointProcessEvents lt
@@ -71,12 +78,14 @@ maybeReconstructedTree :: EpidemicTree -> Either String ReconstructedTree
 maybeReconstructedTree Shoot {} = Left "EpidemicTree is only a Shoot"
 maybeReconstructedTree (Leaf e) =
   case e of
-    IndividualSample {..} -> if indSampSeq
-                             then Right $ RLeaf (Observation e)
-                             else Left "Leaf with non-sequenced event individual sample"
-    PopulationSample {..} -> if popSampSeq
-                             then Right $ RLeaf (Observation e)
-                             else Left "Leaf with non-sequenced event population sample"
+    IndividualSample {..} ->
+      if indSampSeq
+        then Right $ RLeaf (Observation e)
+        else Left "Leaf with non-sequenced event individual sample"
+    PopulationSample {..} ->
+      if popSampSeq
+        then Right $ RLeaf (Observation e)
+        else Left "Leaf with non-sequenced event population sample"
     _ -> Left "Bad leaf in the EpidemicTree"
 maybeReconstructedTree (Branch e@Infection {} lt rt)
   | hasSequencedLeaf lt && hasSequencedLeaf rt = do
@@ -108,7 +117,7 @@ observedEvents epiEvents = do
   let (PointProcessEvents unseqObss) = pointProcessEvents epiTree
   reconTreeEvents <-
     if hasSequencedLeaf epiTree
-      then (liftM reconstructedTreeEvents) $ maybeReconstructedTree epiTree
+      then reconstructedTreeEvents <$> maybeReconstructedTree epiTree
       else Right []
   return $ List.sort . List.nub $ unseqObss ++ reconTreeEvents
 
@@ -120,3 +129,40 @@ reconstructedTreeEvents rt =
       List.sort $
       obs : (reconstructedTreeEvents rtl ++ reconstructedTreeEvents rtr)
     RLeaf obs -> [obs]
+
+-- | Aggregate the sequenced and unsequenced individual level samples
+aggregated :: [TimeInterval] -> [TimeInterval] -> [Observation] -> [Observation]
+aggregated seqAggInts unseqAggInts = List.sort . aggUnsequenced . aggSequenced
+  where
+    aggUnsequenced = _aggregate unseqAggInts False
+    aggSequenced = _aggregate seqAggInts True
+
+-- | Aggregate observations in each of the intervals given the correct
+-- sequencing status.
+_aggregate :: [TimeInterval] -> Bool -> [Observation] -> [Observation]
+_aggregate intervals onlySequenced obs = List.foldl' f obs intervals
+  where
+    f os i = _aggregateInInterval i onlySequenced os
+
+-- | Aggregate all the observations that fall in the interval and have the
+-- correct sequencing status.
+_aggregateInInterval :: TimeInterval -> Bool -> [Observation] -> [Observation]
+_aggregateInInterval interval@TimeInterval {..} onlySequenced obs =
+  let asPopulationSample os absT =
+        Observation $
+        PopulationSample
+          absT
+          (asPeople [indSampPerson ee | Observation ee <- os])
+          onlySequenced
+      (_, aggTime) = timeIntEndPoints
+      toBeAggregated o =
+        case o of
+          Observation (IndividualSample {..}) ->
+            inInterval interval o &&
+            (if onlySequenced
+               then indSampSeq
+               else not indSampSeq)
+          _ -> False
+      (obs2Agg, otherObs) = List.partition toBeAggregated obs
+      newPopSample = asPopulationSample obs2Agg aggTime
+   in newPopSample : otherObs