+++ test/Spec.hs
@@ -239,7 +239,7 @@ eventHandlingTests = do
                 , indSampPerson = p1
                 , indSampSeq = False
                 }
-            , StoppingTime
+            , StoppingTime (AbsoluteTime 6.0)
             ]
           expectedObs =
             [ Observation
@@ -317,6 +317,14 @@ helperFuncTests = do
              isJust (nextTime demoTimed (AbsoluteTime 2.0)) `shouldBe` True
              isJust (nextTime demoTimed (AbsoluteTime 2.1)) `shouldBe` True
              isJust (nextTime demoTimed (AbsoluteTime 10.0)) `shouldBe` True
+           it "the maybeNextTimed function works as expected" $ do
+             let (Just tA) = asTimed [(AbsoluteTime 1, (1.1 :: Double)), (AbsoluteTime 3, 2.3)]
+                 (Just tB) = asTimed [(AbsoluteTime 2, (1 :: Int))]
+             maybeNextTimed tA tB (AbsoluteTime 0.5) == Just (AbsoluteTime 1.0,Left 1.1) `shouldBe` True
+             maybeNextTimed tA tB (AbsoluteTime 1.5) == Just (AbsoluteTime 2.0,Right 1) `shouldBe` True
+             maybeNextTimed tA tB (AbsoluteTime 2.5) == Just (AbsoluteTime 3.0,Left 2.3) `shouldBe` True
+             isNothing (maybeNextTimed tA tB (AbsoluteTime 3.5)) `shouldBe` True
+
     it "shifted times work" $
       let sf =
             fromJust $
@@ -398,6 +406,58 @@ illFormedTreeTest =
           any isReconTreeLeaf simEvents `shouldBe` True
           let (Right oes) = observedEvents simEvents
           (length oes > 1) `shouldBe` True
+
+
+resultAA = demoSampleEvents01
+
+resultAB =
+  [ Infection (AbsoluteTime 1) p1 p2
+  , IndividualSample (AbsoluteTime 3) p1 True
+  , Infection (AbsoluteTime 4) p2 p4
+  , IndividualSample (AbsoluteTime 6) p4 True
+  , PopulationSample (AbsoluteTime 12) (asPeople [p2, p6]) False
+  , IndividualSample (AbsoluteTime 12) p5 True
+  , PopulationSample (AbsoluteTime 17.0) (asPeople []) False
+  ]
+
+resultBA =
+  [ Infection (AbsoluteTime 1) p1 p2
+  , IndividualSample (AbsoluteTime 3) p1 True
+  , Infection (AbsoluteTime 4) p2 p4
+  , IndividualSample (AbsoluteTime 6) p4 True
+  , IndividualSample (AbsoluteTime 8) p2 False
+  , IndividualSample (AbsoluteTime 11) p6 False
+  , PopulationSample (AbsoluteTime 13) (asPeople [p5]) True
+  ]
+
+resultBB =
+  [ Infection (AbsoluteTime 1.0) p1 p2
+  , IndividualSample (AbsoluteTime 3.0) (Person (Identifier 1)) True
+  , Infection (AbsoluteTime 4.0) p2 p4
+  , IndividualSample (AbsoluteTime 6.0) (Person (Identifier 4)) True
+  , PopulationSample (AbsoluteTime 12) (asPeople [p2, p6]) False
+  , PopulationSample (AbsoluteTime 13) (asPeople [p5]) True
+  , PopulationSample (AbsoluteTime 17) (asPeople []) False
+  ]
+
+aggregationTests =
+  describe "Aggregation functionality tests" $ do
+    it "check it does nothing unless it needs to" $
+      let demoObs1 = [Observation ee | ee <- demoFullEvents01]
+          demoObs2 = [Observation ee | ee <- demoSampleEvents01]
+      in do
+        (aggregated [] [] demoObs1) == demoObs1 `shouldBe` True
+        (aggregated [] [] demoObs2) == demoObs2 `shouldBe` True
+    it "check relevant intervals are processed correctly" $
+      let demoObs = [Observation ee | ee <- demoSampleEvents01]
+          demoSeqInts = asConsecutiveIntervals1 [AbsoluteTime 10, AbsoluteTime 13]
+          demoUnseqInts = asConsecutiveIntervals1 [AbsoluteTime 7, AbsoluteTime 12, AbsoluteTime 17]
+      in do
+        aggregated [] [] demoObs == (map Observation resultAA) `shouldBe` True
+        aggregated [] demoUnseqInts demoObs == (map Observation resultAB) `shouldBe` True
+        aggregated demoSeqInts [] demoObs == (map Observation resultBA) `shouldBe` True
+        aggregated demoSeqInts demoUnseqInts demoObs == (map Observation resultBB) `shouldBe` True
+
 
 inhomogeneousBDSTest =
   describe "InhomogeneousBDS module tests" $ do
@@ -677,19 +737,6 @@ newickTests =
           (length demoEvents == 4) `shouldBe` True
           (maybeEpidemicTree demoEvents == maybeEpidemicTree (tail demoEvents)) `shouldBe`
             True
-    -- it "asNewickString works for EpidemicTree" $ do
-    --   let trickyEvents = [
-    --         Infection (AbsoluteTime 0.3) (Person (Identifier 1)) (Person (Identifier 2)),
-    --         Infection (AbsoluteTime 0.4) (Person (Identifier 2)) (Person (Identifier 3)),
-    --         IndividualSample (AbsoluteTime 0.6) (Person (Identifier 3)) True,
-    --         IndividualSample (AbsoluteTime 0.7) (Person (Identifier 1)) True]
-    --   let maybeNewickPair = asNewickString (AbsoluteTime 0, Person (Identifier 1)) =<< maybeEpidemicTree trickyEvents
-    --   let newickTarget = BBuilder.stringUtf8 "(1:0.39999999999999997,(2:Infinity,3:0.19999999999999996):0.10000000000000003):0.3"
-    --   let maybeReconTree = maybeReconstructedTree =<< maybeEpidemicTree trickyEvents
-    --   isJust maybeNewickPair `shouldBe` True
-    --   [IndividualSample (AbsoluteTime 0.6) (Person (Identifier 3)) True, IndividualSample (AbsoluteTime 0.7) (Person (Identifier 1)) True] == snd (fromJust maybeNewickPair) `shouldBe` True
-    --   equalBuilders newickTarget (fst $ fromJust maybeNewickPair) `shouldBe` True
-    --   isJust maybeReconTree `shouldBe` True
         it "asNewickString works for ReconstructedTree" $ do
           isJust
             (asNewickString
@@ -750,3 +797,4 @@ main =
     helperTypeTests
     jsonTests
     newickTests
+    aggregationTests