+++ ChangeLog.md
@@ -1,5 +1,20 @@
 # Changelog for epi-sim
 
+## 0.5.0
+
+- Add absolute times to the extinction and stopping time events to provide a
+  consistent interface.
+- Add the `aggregated` function to help aggregated individual level samples into
+  population level samples. This is tested in `aggregationTests`.
+- Add `TimeStamp` type class to abstract working with types that have an
+  absolute time associated with them.
+- Add `TimeInterval` type for working with intervals of time, there are also
+  some helper functions to make it easier to work with intervals.
+- Add the `maybeNextTimed` helper function and clean up some code in the `Time`
+  module.
+- Extend `ModelParameters` class to have an `eventWeights` to provide a vector
+  of event weights for computing which event actually occurred.
+
 ## 0.4.2
 
 - Include `simulationWithGenIO` and add `scRequireCherry` to the